backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
langchain
elevenlabs
firebase-admin
google-auth
aiofiles
websockets 